refactor: Moved reading time into BlogPost
